fix: fix components replace

+++ 06-build-page/index.js
@@ -36,7 +36,7 @@ async function copyDir(pathFrom, pathTo) {
     });
   }
 
-  const filesFrom = await fsp.readdir(pathFrom,  {withFileTypes: true});
+  const filesFrom = await fsp.readdir(pathFrom, {withFileTypes: true});
   
   filesFrom.forEach(element => {
     const from = path.join(pathFrom, element.name);
@@ -54,19 +54,20 @@ async function copyDir(pathFrom, pathTo) {
   });
   
   let template = await fsp.readFile(path.join(__dirname, 'template.html'), 'utf-8', {withFileTypes: true});
+  const filesFrom = await fsp.readdir(path.join(__dirname, 'components'), {withFileTypes: true});
 
-  const articles = await fsp.readFile(path.join(__dirname, 'components', 'articles.html'), 'utf-8', {withFileTypes: true});
-  const footer = await fsp.readFile(path.join(__dirname, 'components', 'footer.html'), 'utf-8', {withFileTypes: true});
-  const header = await fsp.readFile(path.join(__dirname, 'components', 'header.html'), 'utf-8', {withFileTypes: true});
-  
-  template = template.replaceAll('{{articles}}', articles);
-  template = template.replaceAll('{{footer}}', footer);
-  template = template.replaceAll('{{header}}', header);
-  
-  await fsp.writeFile(path.join(pathTo, 'index.html'), template, (err) => {
-    if (err) throw err;
+  filesFrom.forEach(element => {
+    if (element.isFile()) {
+      fs.readFile(path.join(__dirname, 'components', element.name), 'utf-8', (err, data) => {
+        if (err) throw err;
+        template = template.replaceAll(`{{${element.name.split('.')[0]}}}`, data);
+        fsp.writeFile(path.join(pathTo, 'index.html'), template, (err) => {
+          if (err) throw err;
+        });
+      });
+    }
   });
-  
+
   mergeStyles(path.join(__dirname, 'styles'), path.join(pathTo, 'style.css'));
 
   copyDir(path.join(__dirname, 'assets'), path.join(pathTo, 'assets'));
